requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
openai>=1.30.0
youtube-transcript-api>=0.6.2
feedparser>=6.0.11
python-dotenv>=1.0.1
Pillow>=10.3.0
PyJWT>=2.8.0
requests>=2.32.0
yt-dlp>=2024.1.1
